+++ modules/Admin/Language/ar/Episode.php
@@ -55,6 +55,7 @@ return [
         }',
         'episode' => 'الحلقة',
         'visibility' => 'الظهور',
+        'downloads' => 'Downloads',
         'comments' => 'التعليقات',
         'actions' => 'الإجراءات',
     ],
@@ -85,7 +86,7 @@ return [
             image {cover}
             audio {audio}
             other {media}
-        } file {file_path}. You may manually remove it from your disk.',
+        } file {file_key}. You may manually remove it from your disk.',
         'sameSlugError' => 'An episode with the chosen slug already exists.',
     ],
     'form' => [