+++ modules/Admin/Language/es/Episode.php
@@ -55,6 +55,7 @@ return [
         }',
         'episode' => 'Episodio',
         'visibility' => 'Visibilidad',
+        'downloads' => 'Downloads',
         'comments' => 'Comentarios',
         'actions' => 'Acciones',
     ],
@@ -79,7 +80,7 @@ return [
             audio {audio}
             other {media}
         }.',
-        'deleteFileError' => 'Hubo un problema al tratar de eliminar el archivo {file_path} {type, select,
+        'deleteFileError' => 'Hubo un problema al tratar de eliminar el archivo {file_key} {type, select,
             transcript {de la transcripción}
             chapters {de los episodios}
             image {de la portada}