+++ modules/Admin/Language/fr/Episode.php
@@ -55,6 +55,7 @@ return [
         }',
         'episode' => 'Épisode',
         'visibility' => 'Visibilité',
+        'downloads' => 'Downloads',
         'comments' => 'Commentaires',
         'actions' => 'Actions',
     ],
@@ -85,7 +86,7 @@ return [
             image {couverture}
             audio {audio}
             other {média}
-        } fichier {file_path}. Vous pouvez le supprimer manuellement de votre disque.',
+        } fichier {file_key}. Vous pouvez le supprimer manuellement de votre disque.',
         'sameSlugError' => 'Il existe déjà un épisode avec le slug choisi.',
     ],
     'form' => [